backend/pyproject.toml
[project]
name = "server-provision-backend"
version = "0.1.0"
description = "Flask Form API for the Smart Server Provisioning Assistant"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "pymongo>=4.7",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
